move iOS to SkiaMultiplatformSample
