+++ samples/SkiaMultiplatformSample/build.gradle.kts
@@ -14,6 +14,7 @@ buildscript {
 
 plugins {
     kotlin("multiplatform") version "1.6.10"
+    id("org.jetbrains.gradle.apple.applePlugin") version "222.849-0.15.1"
 }
 
 val coroutinesVersion = "1.5.2"
@@ -72,6 +73,27 @@ kotlin {
 
         targets.add(iosX64())
         targets.add(iosArm64())
+
+        ios {
+            binaries {
+                framework {
+                    baseName = "shared"
+                    freeCompilerArgs += listOf(
+                        "-linker-option", "-framework", "-linker-option", "Metal",
+                        "-linker-option", "-framework", "-linker-option", "CoreText",
+                        "-linker-option", "-framework", "-linker-option", "CoreGraphics"
+                    )
+                }
+//            executable {
+//                entryPoint = "me.user.shared.main"
+//                freeCompilerArgs += listOf(
+//                    "-linker-option", "-framework", "-linker-option", "Metal",
+//                    "-linker-option", "-framework", "-linker-option", "CoreText",
+//                    "-linker-option", "-framework", "-linker-option", "CoreGraphics"
+//                )
+//            }
+            }
+        }
     }
 
     jvm("awt") {
@@ -149,7 +171,7 @@ kotlin {
                 }
                 else -> throw GradleException("Host OS is not supported")
             }
-            val iosMain by creating {
+            val iosMain by getting {
                 dependsOn(darwinMain)
             }
             val iosX64Main by getting {
@@ -284,4 +306,14 @@ if (hostOs == "macos") {
 
 rootProject.plugins.withType<org.jetbrains.kotlin.gradle.targets.js.nodejs.NodeJsRootPlugin> {
     rootProject.the<org.jetbrains.kotlin.gradle.targets.js.nodejs.NodeJsRootExtension>().nodeVersion = "16.0.0"
+}
+
+apple {
+    iosApp {
+        productName = "SkikoAppCode"
+        sceneDelegateClass = "SceneDelegate"
+        dependencies {
+            implementation(project(":"))
+        }
+    }
 }
