+++ samples/SkiaMultiplatformSample/gradle.properties
@@ -1,3 +1,4 @@
 kotlin.code.style=official
 kotlin.mpp.enableGranularSourceSetsMetadata=true
 kotlin.native.enableDependencyPropagation=false
+org.gradle.jvmargs=-Xmx3G -XX:MaxMetaspaceSize=512m
