+++ samples/SkiaMultiplatformSample/settings.gradle.kts
@@ -8,3 +8,8 @@ pluginManagement {
 }
 rootProject.name = "SkiaMultiplatformSample"
 
+includeBuild("../../skiko") {
+    dependencySubstitution {
+        substitute(module("org.jetbrains.skiko:skiko")).using(project(":"))
+    }
+}
