+++ samples/SkiaMultiplatformSample/src/iosAppMain/apple/AppDelegate.swift
