+++ samples/SkiaMultiplatformSample/src/iosAppMain/apple/Assets.xcassets/AppIcon.appiconset/Contents.json
