+++ samples/SkiaMultiplatformSample/src/iosAppMain/apple/Assets.xcassets/Contents.json
