+++ samples/SkiaMultiplatformSample/src/iosAppMain/apple/SceneDelegate.swift
