--- samples/ios-appcode/README.md
@@ -1,9 +1,0 @@
-## How to run:
- - Install KMM plugin for AppCode
- - In AppCode open samples/ios-appcode/build.gradle.kts (File -> Open).
-Choose "Open as Project".
-![import-build-gradle-project.png](import-build-gradle-project.png)
- - Set target device and Run
-![ios-run-in-appcode.png](ios-run-in-appcode.png)
- - Now you may use breakpoints in common and native Kotlin code
-
