--- samples/ios-appcode/gradle.properties
@@ -1,5 +1,0 @@
-org.gradle.jvmargs=-Xmx3G -XX:MaxMetaspaceSize=512m
-kotlin.mpp.enableGranularSourceSetsMetadata=true
-kotlin.native.enableDependencyPropagation=false
-
-
