--- samples/ios-appcode/iosApp/build.gradle.kts
@@ -1,60 +1,0 @@
-plugins {
-    id("org.jetbrains.gradle.apple.applePlugin") version "222.849-0.15.1"
-    kotlin("multiplatform") version "1.6.20"
-}
-
-repositories {
-    google()
-    jcenter()
-}
-val versionSkiko = "0.0.0-SNAPSHOT"
-val coroutinesVersion = "1.5.2"
-
-kotlin {
-    ios {
-        binaries {
-            framework {
-                baseName = "shared"
-                freeCompilerArgs += listOf(
-                    "-linker-option", "-framework", "-linker-option", "Metal",
-                    "-linker-option", "-framework", "-linker-option", "CoreText",
-                    "-linker-option", "-framework", "-linker-option", "CoreGraphics"
-                )
-            }
-//            executable {
-//                entryPoint = "me.user.shared.main"
-//                freeCompilerArgs += listOf(
-//                    "-linker-option", "-framework", "-linker-option", "Metal",
-//                    "-linker-option", "-framework", "-linker-option", "CoreText",
-//                    "-linker-option", "-framework", "-linker-option", "CoreGraphics"
-//                )
-//            }
-        }
-    }
-    sourceSets {
-        val commonMain by getting {
-            dependencies {
-                implementation("org.jetbrains.skiko:skiko:$versionSkiko")
-                implementation("org.jetbrains.kotlinx:kotlinx-coroutines-core:$coroutinesVersion")
-            }
-        }
-        val commonTest by getting {
-            dependencies {
-                implementation(kotlin("test"))
-            }
-        }
-        val iosMain by getting
-        val iosTest by getting
-    }
-}
-
-
-apple {
-    iosApp {
-        productName = "SkikoAppCode"
-        sceneDelegateClass = "SceneDelegate"
-        dependencies {
-            implementation(project(":iosApp"))
-        }
-    }
-}
