--- samples/ios-appcode/iosApp/src/iosMain/kotlin/org/jetbrains/skiko/sample/SwiftHelper.kt
@@ -1,7 +1,0 @@
-package org.jetbrains.skiko.sample
-
-import org.jetbrains.skiko.sample.getSkikoViewContoller
-
-class SwiftHelper {
-    fun getViewController() = getSkikoViewContoller()
-}
