--- samples/ios-appcode/settings.gradle.kts
@@ -1,6 +1,0 @@
-include("iosApp")
-includeBuild("../../skiko") {
-    dependencySubstitution {
-        substitute(module("org.jetbrains.skiko:skiko")).using(project(":"))
-    }
-}
